moved auth fragments
