+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name="com.aa.draftt.auth.views.AuthActivity">
+        <activity android:name="com.aa.draftt.views.auth.AuthActivity">
         </activity>
         <activity android:name="com.aa.draftt.views.HomeActivity">
         </activity>
